requirements.txt
torch
torchvision
os
numpy
pickle
pandas
time
copy
matplotlib
Pillow
h5py